KAN-65 Refactor: 질문 페이지 컴포넌트/애니메이션 모듈화
